git.status now accepts an array of paths to be used for the filtering.
It already expected an array but would only filter using the first item.

Filtering is now case sensitive — it seems wrong to do case-insenstive filtering considering that git is not case-insensitive, nor is all file systems (for HFS+ it is the default setting, but not required).

While filtering it no longer updates path/display of the items as It is unclear to me what this is good for.