Word wrap user’s commit message in result window.
Also use htmlize_attr instead of htmlize. The latter is not for use with <pre> as it will convert whitespace sequences to use non-breaking spaces and newlines to ‘<br>\n’ (which was resulting in doubling the line breaks).

Semantically htmlize_attr is for escaping attribute value content, but in practice it is simply entity-encoding the content.